tool(webpack): Extract js files discovery
This will allow to reuse the logic in Polylang Pro
