+++ webpack.config.js
@@ -1,11 +1,22 @@
 /**
+ * @package Polylang
+ */
+
+ /**
  * External dependencies
  */
+
 const path = require( 'path' );
 const glob = require( 'glob' ).sync;
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const CssMinimizerPlugin = require('css-minimizer-webpack-plugin');
-const { CleanWebpackPlugin } = require('clean-webpack-plugin');
+const MiniCssExtractPlugin = require( 'mini-css-extract-plugin' );
+const CssMinimizerPlugin = require( 'css-minimizer-webpack-plugin' );
+const { CleanWebpackPlugin } = require( 'clean-webpack-plugin' );
+
+/**
+ * Internal dependencies
+ */
+
+ const getJsFileNamesEntries = require( './webpack/FileNames' );
 
 function configureWebpack( options ){
 	const mode = options.mode;
@@ -17,8 +28,10 @@ function configureWebpack( options ){
 	const jsFoldersToIgnore = [
 		'node_modules/**/*.js',
 		'vendor/**/*.js',
-		'tmp/**'
+		'tmp/**',
+		'webpack/*.js'
 	];
+
 	const jsFileNamesToIgnore = [
 		'*.config.js',
 		'**/*.src.js',
@@ -26,55 +39,17 @@ function configureWebpack( options ){
 		'**/*.dep.js',
 	];
 	
-	const jsSourceFileNames = glob( '**/*.src.js', { 'ignore': jsFoldersToIgnore } ).map( filename => `./${ filename }` );
-	console.log( 'js files to build:', jsSourceFileNames );
+	const jsFileNamesEntries = getJsFileNamesEntries( jsFoldersToIgnore, jsFileNamesToIgnore );
 
-	const jsFileNames = glob( 
-		'**/*.js', 
-		{ 
-			'ignore': [
-				...jsFoldersToIgnore, 
-				...jsFileNamesToIgnore, 
-				// Glob ignore pathes cannot use the '.' special character
-				...jsSourceFileNames.map( filename => computeBuildFilename( filename ).substr( 2 ) )
-			]
-		} 
-	).map( filename => `./${ filename }` );
-	console.log( 'js files to minify:', jsFileNames );
+	const cssFileNamesToIgnore = [
+		'node_modules/**/*.css',
+		'vendor/**/*.css',
+		'tmp/**',
+		'**/*.min.css'
+	]
 
-	const cssFileNames = glob( '**/*.css', { 'ignore': commonFileNamesToIgnore } ).map( filename => `./${ filename }`);
+	const cssFileNames = glob( '**/*.css', { 'ignore': cssFileNamesToIgnore } ).map( filename => `./${ filename }`);
 	console.log( 'css files to minify:', cssFileNames );
-
-	function computeBuildFilename( filename, suffix ) {
-		const nameWithoutSuffix = path.parse( filename ).name.split( '.' )[0];		
-		suffix = suffix ? '.' + suffix : '';
-
-		return `${ path.parse( filename ).dir }/${ nameWithoutSuffix + suffix }.js`;
-	}
-
-	// Prepare webpack configuration to minify js files to source folder as target folder and suffix file name with .min.js extension.
-	function mapJsFiles( jsFileNames, minimize = false ) {
-		return jsFileNames.map( ( filename ) => {
-			const entry = {};
-			entry[ path.parse( filename ).name ] = filename;
-			const output = {
-				filename: computeBuildFilename( filename, minimize ? 'min' : '' ),
-				path: path.resolve( __dirname ), // Output folder as project root to put files in the same folder as source files.
-				iife: false, // Avoid Webpack to wrap files into a IIFE which is not needed for this kind of javascript files.
-			}
-			const config = {
-				entry: entry,
-				output: output,
-				optimization: { minimize: minimize }
-			};
-			return config;
-		},
-		{});
-	}
-	const jsFileNamesEntries = [
-		...mapJsFiles( [ ...jsFileNames, ...jsSourceFileNames ], true ),
-		...mapJsFiles( jsSourceFileNames )
-	];
 
 	// Prepare webpack configuration to minify css files to source folder as target folder and suffix file name with .min.js extension.
 	const cssFileNamesEntries = cssFileNames.map( ( filename ) => {
@@ -138,3 +113,4 @@ function configureWebpack( options ){
 module.exports = ( env, options ) => {
 	return configureWebpack( options );
 }
+
